# internal/testapiserver/testdata/openapi.yaml
...
openapi: 3.0.3
info:
  title: quirkserver
  description: the deliberately-misbehaving stub API the toolkit rehearses against
  version: "1.0"
paths:
...
  /monitors:
    get:
      operationId: listMonitors
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Monitor'
    post:
      operationId: createMonitor
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MonitorCreate'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Monitor'
...
  /assignments:
    get:
      operationId: listAssignments
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Assignment'
    post:
      operationId: createAssignment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentCreate'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
...
  /streams:
    get:
      operationId: listStreams
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stream'
    post:
      operationId: createStream
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StreamCreate'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stream'
...
components:
  schemas:
...
    MonitorCreate:
      type: object
      # Only kind is declared required. interval is really required too and the
      # per-kind fields are gated by kind's value, but the document says none of
      # that — the audit finds it.
      required: [kind]
      properties:
        kind:
          type: string
          enum: [ping, web, dns]
        name:
          type: string
        interval:
          type: integer
        target_host:
          type: string
        domain:
          type: string
        dnssec:
          type: boolean
        web:
          type: object
          properties:
            url:
              type: string
            method:
              type: string
    Monitor:
      allOf:
        - $ref: '#/components/schemas/MonitorCreate'
        - type: object
          properties:
            id:
              type: string
              readOnly: true
    AssignmentCreate:
      type: object
      required: [name, agent_id]
      properties:
        name:
          type: string
        agent_id:
          type: string
    Assignment:
      allOf:
        - $ref: '#/components/schemas/AssignmentCreate'
        - type: object
          properties:
            id:
              type: string
              readOnly: true
...
    StreamCreate:
      type: object
      # format and mode are declared plain required enums. That format
      # value-gates which mode is valid — and that the API says so in vendor
      # prose, not the monitor's parseable grammar — is what the audit finds by
      # value-cycling.
      required: [name, format, mode]
      properties:
        name:
          type: string
        format:
          type: string
          enum: [avro, json, locked]
        mode:
          type: string
          enum: [batch, streaming]
    Stream:
      allOf:
        - $ref: '#/components/schemas/StreamCreate'
        - type: object
          properties:
            id:
              type: string
              readOnly: true